:fix: #91 : move CoreStaticFileHandler to standard api ressources.
Please be carefull when merging, this breaks existing projects that
implement a CoreStaticFileHandle, as the qual_name changed and
therefore the import statements need to be fixed.

Also, the standard user now can view all applications served by such a
CoreStaticFileHandler, whether he has access to the api-endpoints of
that project or not.